Improve performance generating thumbnails
It turns out there's a specific thumbnail() function that got added in
image v0.19.0, that is almost twice as fast as the resize() function.

However, it has a bug that causes it to panic when the output image
is larger than the input. It's fixed in the latest code, but the
fix is unreleased, so use resize() for that case for now.